"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ assignments/m2/Assignment-3/Assignment-3/Solution.java
@@ -9,7 +9,7 @@ public class Solution {
      *
      * @return     it returns the value
      */
-    public static int power(int base,int exponent){
+    public static long power(int base,int exponent){
        
         {
             if (exponent > 0)
